// website/src/app/testing/tests-implementation/speed/enums/speed-test-state.enum.ts
export enum SpeedTestStateEnum {
    ERROR = "ERROR",
    READY = "READY",
    WAITING = "WAITING",
    STARTING = "STARTING",
    CONNECTED = "CONNECTED",
    TOKEN = "TOKEN",
    TOKEN_OK = "TOKEN_OK",
    DOWN_PRE = "DOWN_PRE",
    DOWN_PRE_OK = "DOWN_PRE_OK",
    PING = "PING",
    PING_OK = "PING_OK",
    DOWN = "DOWN",
    DOWN_OK = "DOWN_OK",
    UP_PRE = "UP_PRE",
    UP_PRE_OK = "UP_PRE_OK",
    UP = "UP",
    UP_OK = "UP_OK",
    SUBMIT = "SUBMIT",
    SUBMIT_OK = "SUBMIT_OK",
    QOS = "QOS",
    QOS_OK = "QOS_OK",
    QOS_SUBMIT = "QOS_SUBMIT",
    QOS_SUBMIT_OK = "QOS_SUBMIT_OK",
    COMPLETE = "COMPLETE",
}